Switch to flake with Python+poetry+virtualenv
I was having issues with jupyenv, where I couldn't add `ipympl` as a
dependency which I need for the course because it had a version
conflict with another dependency as both needed  incompatible versions
of `ipython-genutils`.

I tried the poetry way indicated in the documentation, where you can
specify your own `pyproject.toml` and `poetry.lock` to the kernel
configuration and have it set up with those dependencies. But for some
reason whenever I tried to build, it'd always fail to find the
project/files. I tried looking at the repo code, and issues, and I
didn't find anything indicating there were any breaking changes, so
not sure why it's not working.

Regardless, I'm not interested in delving into the jupyenv Nix code to
try to figure it out. With this new flake I've switched to just having
Python/virtualenv/poetry install a venv, and then manage the
dependencies in user land with poetry using the `pyproject.toml`,
which so far seems to work.